Add a profiler based on async-profiler
async-profiler is "a low overhead sampling profiler for Java that
does not suffer from Safepoint bias problem. It features HotSpot-specific
APIs to collect stack traces and to track memory allocations."

An option parser handles the full range of current features of
async-profiler. Some measure of future proofing is provided with
the option `rawCommand` that exposes the underlying command API
of the profiler; this could be used to set newly added
configuration options without requiring a new release of JMH.

This integration controls the profiler as in internal profiler via
its Java API. Samples collected during the warmup period are
discarded.

async-profiler must be installed separately by the user. It can be
included in LD_LIBRARY_PATH, -Djava.library.path, or via an option
to the profiler.

At the conclusion of the final measurement iterations, the results
are rendered to one or more of the output formats supported by
the profiler (currently: flamegraph,tree,text,jfr). By default
both forward and reverse flamegraphs are generated; this behaviour
is configurable. If the text output is among the output formats,
it is also written to console.

All of these files are written to a per-trial directory that is
located within the user specified output directory. If the user
does not specify an output directory, one will be created automatically.